requirements.txt
Django==3.0.8
django-filter==2.3.0
requests==2.23.0
